added $filename and if exists

+++ allen/web-forms/viewwebfornscvdatabase.php
@@ -1,19 +1,24 @@
 <?php
 
-//open database for reading
-$database = fopen("database.csv", "r");
+$filename = "database.csv";
 
-//Read each line and print to screen
-while (!feof($database))
+if (file_exists($filename))
 {
-	$line = fgets($database);
-	$data = str_getcsv($line);
-	
-	echo "First Name: " . $data[0] . "<br/>";
-	echo "Last Name: " . $data[1] . "<br/>";
-	echo "Favorite Color: " . $data[2] . "<br/>";
-	echo "Password: " . $data[3] . "<br/>";
-}
+	//open database for reading
+	$database = fopen($filename, "r");
 
-//Be Nice! - Close the database
-fclose($database);
+	//Read each line and print to screen
+	while (!feof($database))
+	{
+		$line = fgets($database);
+		$data = str_getcsv($line);
+		
+		echo "First Name: " . $data[0] . "<br/>";
+		echo "Last Name: " . $data[1] . "<br/>";
+		echo "Favorite Color: " . $data[2] . "<br/>";
+		echo "Password: " . $data[3] . "<br/>";
+	}
+
+	//Be Nice! - Close the database
+	fclose($database);
+}
